pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Book Store API"
authors = ["Pedro Barbas <pedrodrlvbarbas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0"
djangorestframework = "^3.15.2"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.4"
jinja2 = "^3.1.4"
flake8 = "^7.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
factory-boy = "^3.3.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
